Adding CSV output

+++ backup.sh
@@ -13,6 +13,7 @@ DESTINATION="."
 NAME="%Y%m%d-%H%M%S.sql"
 DB=""
 THEN=""
+OUTPUT="sql"
 
 # Dynamic vars
 cmdname=$(basename "${0}")
@@ -37,17 +38,18 @@ Usage:
     -w password     Password for user, defaults to empty
     -W path         Same as -w, but read content of password from file instead
     -d destination  Directory where to place (and rotate) backups.
-    -n basename     Basename for file to create, date-tags allowed, defaults to: %Y%m%d-%H%M%S.sql
+    -n basename     Basename for file/dir to create, date-tags allowed, defaults to: %Y%m%d-%H%M%S.sql
     -k keep         Number of backups to keep, defaults to empty, meaning keep all backups
     -b database     Name of database to backup, defaults to empty, meaning all databases
     -t command      Command to execute once done, path to backup will be passed as an argument.
+    -o output       Output type: sql or csv (only tables content).
 USAGE
   exit "$exitcode"
 }
 
 
 # Parse options 
-while getopts ":k:h:p:u:w:d:n:b:vt:W:" opt; do
+while getopts ":k:h:p:u:w:d:n:b:vt:W:o:" opt; do
     case $opt in
         k)
             KEEP="$OPTARG"
@@ -85,6 +87,9 @@ while getopts ":k:h:p:u:w:d:n:b:vt:W:" opt; do
         t)
             THEN="$OPTARG"
             ;;
+        o)
+            OUTPUT="$OPTARG"
+            ;;
         \?)
             echo "Invalid option: $opt" >& 2
             usage 1
@@ -97,6 +102,13 @@ while getopts ":k:h:p:u:w:d:n:b:vt:W:" opt; do
 done
 shift $((OPTIND-1))
 
+# Returns absolute path for file, from: https://stackoverflow.com/a/20500246
+abspath() {
+    cd "$(dirname "$1")"
+    printf "%s/%s\n" "$(pwd)" "$(basename "$1")"
+    cd "$OLDPWD"
+}
+
 log() {
     local txt=$1
 
@@ -105,24 +117,52 @@ log() {
     fi
 }
 
+csv_dump() {
+    local db=$1
+
+    log "Starting $OUTPUT backup of database $db to $FILE"
+    mkdir -p $FILE/$db
+    TABLES=$(echo "SELECT table_schema || '.' || table_name FROM information_schema.tables WHERE table_type = 'BASE TABLE' AND table_schema NOT IN ('pg_catalog', 'information_schema');" | psql -h $HOST -p $PORT -U $USER -d $db -q -t)
+    for table in $TABLES; do
+        DST=$FILE/$db/${table}.csv
+        log "Dumping $table to $DST"
+        echo "COPY $table TO STDOUT WITH CSV HEADER;" | psql -h $HOST -p $PORT -U $USER -d $db > $DST
+    done
+}
+
 if [ -n "$PASSWORDFILE" ]; then
     PASSWORD=$(cat $PASSWORDFILE)
 fi
 export PGPASSWORD=$PASSWORD
 FILE=$(date +$NAME)
-if [ -z "${DB}" ]; then
-    log "Starting backup of all databases to $FILE"
-    CMD="pg_dumpall -h $HOST -p $PORT -U $USER -w -f ${DESTINATION}/$FILE"
-else 
-    log "Starting backup of database $DB to $FILE"
-    CMD="pg_dump -h $HOST -p $PORT -U $USER -w -f ${DESTINATION}/$FILE $DB"
-fi
 
-if $CMD; then
-    log "Backup done"
-else
-    echo "Could not create backup!" >& 2
-    rm -rf ${DESTINATION}/$FILE
+if [ "$OUTPUT" = "sql" ]; then
+    if [ -z "${DB}" ]; then
+        log "Starting $OUTPUT backup of all databases to $FILE"
+        CMD="pg_dumpall -h $HOST -p $PORT -U $USER -w -f ${DESTINATION}/$FILE"
+    else 
+        log "Starting $OUTPUT backup of database $DB to $FILE"
+        CMD="pg_dump -h $HOST -p $PORT -U $USER -w -f ${DESTINATION}/$FILE $DB"
+    fi
+
+    if $CMD; then
+        log "Backup done"
+    else
+        echo "Could not create backup!" >& 2
+        rm -rf ${DESTINATION}/$FILE
+    fi
+elif [ "$OUTPUT" = "csv" ]; then
+    if [ -z "${DB}" ]; then
+        log "Starting $OUTPUT backup of all databases to $FILE"
+        DBS=$(psql -h $HOST -p $PORT -U $USER -l -t|cut -d "|" -f 1| sed "s/^\s*//g")
+        for db in $DBS; do
+            if [ -n "$db" ]; then
+                csv_dump $db
+            fi
+        done
+    else
+        csv_dump $DB
+    fi
 fi
 
 if [ -n "${KEEP}" ]; then
